Functions now externalized. Pixel selection criteria is now controlled by a new function shared by both the injector and reverser.
